Bugfix: Naming cookies on Firebase CDN (#12)
Previously the cookie was never read properly on first page load in a
auction.
This meant that you couldn't tell which bidder/color you were. 

This was a problem only in production. 

According to https://firebase.google.com/docs/hosting/manage-cache the
only cookie that is loaded correctly on CDN is `__session`.
This PR renames the cookie accordingly.