Remove the priority tiebreaker (#23)
This change removes the second tiebreaker. Instead a random winner is
drawn if two bidders bid the same AND have the same amount of points
left.
The tiebreaker was hard to explain, rarely came up, but still gave an
advantage to the person who started the auction. So I chose to remove
it.